Create and include version information bin resource
Added pre- and post-build events to the .dproj file to generate and
compile a temporary resource source file. The temporary file is created
by calling BogBat to replace the templates in
src/VersionInfo.rc.template with the values in src/VERSION. The
resulting file is then compiled using BRCC32 and then deleted.

The .dpr file was modified to link in VersionInfo.res file generated by
the build events.